materials repo tests added

+++ src/RepositoryServices/materialRepositoryService.ts
@@ -1,5 +1,5 @@
 import RepositoryService from "../interfaces/service.interface";
-import {DeleteResult, getRepository} from "typeorm";
+import {DeleteResult, getRepository, UpdateResult} from "typeorm";
 import Material from "../Models/Materials/material.entity";
 import MaterialNotFoundExceptionn from "../Exceptions/MaterialNotFoundException";
 import CreateMaterialDto from "../Models/Materials/material.dto";
@@ -21,19 +21,14 @@ class MaterialService implements RepositoryService{
     }
     public async findOneMaterialByMaterialCode(materialCode:string):Promise<Material>{
         const foundMaterial:Material=await this.repository.findOne({materialCode:materialCode});
-        if(!foundMaterial){
-            throw new MaterialNotFoundExceptionn(materialCode);
-        }
+
         return foundMaterial;
 
 
     }
     public async findOneMaterialByMaterialName(materialName:string):Promise<Material>{
         const foundMaterial:Material=await this.repository.findOne({materialName:materialName});
-        if(!foundMaterial){
-            throw new MaterialNotFoundExceptionn(materialName);
-        }
-        return foundMaterial;
+       return foundMaterial;
 
 
     }
@@ -47,12 +42,17 @@ class MaterialService implements RepositoryService{
     public async addOneMaterial(createMaterialDto:CreateMaterialDto):Promise<Material>{
         const materialWithThisCodeInDatabase=await this.findOneMaterialByMaterialCode(createMaterialDto.materialCode);
         const materialWithThisNameInDatabase=await this.findOneMaterialByMaterialName(createMaterialDto.materialName);
-        if(materialWithThisCodeInDatabase){
-            throw new MaterialAlreadyExistsException(createMaterialDto.materialCode);
+
+        if(materialWithThisCodeInDatabase&&materialWithThisNameInDatabase){
+            throw new MaterialAlreadyExistsException(createMaterialDto.materialCode,createMaterialDto.materialName);
+        }
+        else if(materialWithThisCodeInDatabase){
+            throw new MaterialAlreadyExistsException(createMaterialDto.materialCode,null);
         }
         else if(materialWithThisNameInDatabase){
-            throw new MaterialAlreadyExistsException(createMaterialDto.materialName);
+            throw new MaterialAlreadyExistsException(null,createMaterialDto.materialName);
         }
+
         const materialTosave:Material={
             ...createMaterialDto
         };
@@ -61,22 +61,47 @@ class MaterialService implements RepositoryService{
 
     }
     public async updateMaterialById(id:string,createMaterialDto:CreateMaterialDto):Promise<Material>{
-        const materialWithThisCodeinDatabase=await this.findOneMaterialByMaterialCode(createMaterialDto.materialCode);
-        const materialWithThisNameinDatabase=await this.findOneMaterialByMaterialName(createMaterialDto.materialName);
-        // do not allow to update if other material with this code or name already exist and throw exception
-        if(materialWithThisCodeinDatabase&&materialWithThisCodeinDatabase.id!==Number(id)){
-            throw new MaterialAlreadyExistsException(createMaterialDto.materialCode);
+        const idOfExistingUser:boolean=await this.findOneMaterialById(id)!==null;
+        if(idOfExistingUser){
+            const materialWithThisCodeInDatabase=await this.findOneMaterialByMaterialCode(createMaterialDto.materialCode);
+            const materialWithThisNameInDatabase=await this.findOneMaterialByMaterialName(createMaterialDto.materialName);
+            // do not allow to update if other material with this code or name already exist and throw exception
+            if((materialWithThisCodeInDatabase&&materialWithThisNameInDatabase)){
+                if((materialWithThisCodeInDatabase.id!==Number(id)||materialWithThisNameInDatabase.id!==Number(id))){
+                    throw new MaterialAlreadyExistsException(createMaterialDto.materialCode,createMaterialDto.materialName);
+                }
+
+            }
+            else if(materialWithThisCodeInDatabase){
+                if(materialWithThisCodeInDatabase.id!==Number(id)){
+                    throw new MaterialAlreadyExistsException(createMaterialDto.materialCode,null);
+                }
+
+            }
+            else if(materialWithThisNameInDatabase){
+                if(materialWithThisNameInDatabase.id!==Number(id)){
+                    throw new MaterialAlreadyExistsException(null,createMaterialDto.materialName);
+                }
+
+            }
+            const updateResult:UpdateResult=await this.repository.update(id,createMaterialDto);
+            if(updateResult.affected===1){
+                const updatedMaterial:Material=await this.findOneMaterialById(id);
+                return updatedMaterial;
+            }
+
         }
-        else if(materialWithThisNameinDatabase&&materialWithThisNameinDatabase.id!==Number(id)){
-            throw new MaterialAlreadyExistsException(createMaterialDto.materialName);
-        }
-        const updatedMaterial:Material=await this.repository.update(id,createMaterialDto);
-        return updatedMaterial;
+
+
 
     }
-    public async deleteMaterialById(id:string):DeleteResult{
-       const deleteResult:DeleteResult= await this.repository.delete(id);
-       return deleteResult;
+    public async deleteMaterialById(id:string):Promise<DeleteResult>{
+       const idOfExistingUser:boolean=await this.findOneMaterialById(id)!==null;
+       if(idOfExistingUser){
+           const deleteResult:DeleteResult= await this.repository.delete(id);
+           return deleteResult;
+       }
+
